pyproject.toml
[tool.poetry]
name = "biden nephew"
version = "0.1.0"
description = ""
authors = ["zacchaeus <zacchaeuschok2014@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
pyqt6 = "^6.8.0"
pyautogui = "^0.9.54"
requests = "^2.32.3"
anthropic = "^0.43.1"
python-dotenv = "^1.0.1"
pillow = "^11.1.0"
numpy = "^2.2.1"
qtawesome = "^1.3.1"
pre-commit = "^4.0.1"
pynput = "^1.7.7"
python-xlib = { version = "^0.33", platform = "linux" }
pyobjc-framework-Quartz = { version = "^11.0", platform = "darwin" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
